+++ CHANGELOG.md
@@ -7,7 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/iamdefinitelyahuman/brownie-token-tester)
 
+## [0.2.2](https://github.com/iamdefinitelyahuman/brownie-token-tester/tree/v0.2.2) - 2021-06-01
+
+### Changed
+- Generalize minting logic for Synthetix synths
+
+## [0.2.1](https://github.com/iamdefinitelyahuman/brownie-token-tester/tree/v0.2.1) - 2021-05-26
+
+### Changed
+- Use Vyper `v0.2.12`
+
+### Fixed
+- Query SNX exchanger address from address resolver
+
 ## [0.2.0](https://github.com/iamdefinitelyahuman/brownie-token-tester/tree/v0.2.0) - 2021-03-28
+
 ### Added
 
 - Custom minting logic for: USDC, DAI, USDT, EURS, sEURS, sETH, sUSD, ankrETH, rETH